[1101] refactor have_providers matcher
DRY'd custom matchers for provider in the same way as course
