+++ spec/spec_helper.rb
@@ -3,6 +3,7 @@ require 'rspec-json_matchers'
 require 'pundit/rspec'
 require 'jsonapi/rspec'
 require 'support/matchers/have_courses'
+require 'support/matchers/have_providers'
 
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
